# Optimize acquiring and releasing read locks when one is already held