--- a/src/realm/version_id.hpp
+++ b/src/realm/version_id.hpp
@@ -34,36 +34,34 @@ struct VersionID {
     version_type version = std::numeric_limits<version_type>::max();
     uint_fast32_t index = 0;
 
-    VersionID()
-    {
-    }
-    VersionID(version_type initial_version, uint_fast32_t initial_index)
+    constexpr VersionID() = default;
+    constexpr VersionID(version_type initial_version, uint_fast32_t initial_index) noexcept
     {
         version = initial_version;
         index = initial_index;
     }
 
-    bool operator==(const VersionID& other) const
+    constexpr bool operator==(const VersionID& other) const noexcept
     {
         return version == other.version;
     }
-    bool operator!=(const VersionID& other) const
+    constexpr bool operator!=(const VersionID& other) const noexcept
     {
         return version != other.version;
     }
-    bool operator<(const VersionID& other) const
+    constexpr bool operator<(const VersionID& other) const noexcept
     {
         return version < other.version;
     }
-    bool operator<=(const VersionID& other) const
+    constexpr bool operator<=(const VersionID& other) const noexcept
     {
         return version <= other.version;
     }
-    bool operator>(const VersionID& other) const
+    constexpr bool operator>(const VersionID& other) const noexcept
     {
         return version > other.version;
     }
-    bool operator>=(const VersionID& other) const
+    constexpr bool operator>=(const VersionID& other) const noexcept
     {
         return version >= other.version;
     }